chore: 🤖 Prelim Tests
Commit for creating branch static_ray_cast
